Change Contact Us configuration chnage and lat and lan

+++ Application/DTOs/ContactUses/ContactUsCreateDTO.cs
@@ -13,8 +13,8 @@ namespace Application.DTOs.ContactUs
         public required string Address { get; set; }
         [MaxLength(500, ErrorMessage = "آدرس  اجباری است طول 500 کاراکتر است ")]
         public required string Email { get; set; }
-        public float Lan { get; set; }
-        public float Lat { get; set; }
+        public double Lan { get; set; }
+        public double Lat { get; set; }
 
     }
 }
